+++ ChoiceSosyalHizmet.DAL/VM/VMPersonel.cs
@@ -8,7 +8,7 @@ namespace ChoiceSosyalHizmet.DAL.VM
 {
    public class VMPersonel
     {
-        public int PersonelID { get; set; }
+        public long PersonelID { get; set; }
         public string AdiSoyadi { get; set; }
         public string TC { get; set; }
     }
